[TS-301] Refactor: card-detail-star 이미지 이름 변경
